+++ src/vigilant/driver/__init__.py
@@ -2,6 +2,9 @@ import os
 import yaml
 from pathlib import Path
 
+RED = "\033[31m"
+RESET = "\033[0m"
+
 def set_env_variables_from_dict(data):
     for key, value in data.items():
         os.environ[key] = str(value)
@@ -20,5 +23,5 @@ if Path(CONFIG_YAML_FILE).exists():
     print(f"Setting configuration from {CONFIG_YAML_FILE} file.")
     set_env_variables_from_yaml(CONFIG_YAML_FILE)
 else:
-    print(f"Could not find `{CONFIG_YAML_FILE}` configuration file.")
-    print("If you don't want to use the yaml configuration file, ensure you provide all configuration data as environment variables.")
+    print(f"{RED}Could not find `{CONFIG_YAML_FILE}` configuration file.{RESET}")
+    print(f"{RED}If you don't want to use the yaml configuration file, ensure you provide all configuration data as environment variables.{RESET}")